Redirect embedded apps to the provided return_to if it's a fully-formed URL

+++ app/controllers/shopify_app/callback_controller.rb
@@ -69,13 +69,18 @@ module ShopifyApp
     end
 
     def redirect_to_app
-      if ShopifyAPI::Context.embedded?
-        return_to = "#{decoded_host}#{session.delete(:return_to)}"
-        return_to = ShopifyApp.configuration.root_url if deduced_phishing_attack?
-        redirect_to(return_to, allow_other_host: true)
+      if ShopifyAPI::Context.embedded? && !fully_formed_url?(session[:return_to])
+          return_to = "#{decoded_host}#{session.delete(:return_to)}"
+          return_to = ShopifyApp.configuration.root_url if deduced_phishing_attack?
+          redirect_to(return_to, allow_other_host: true)
       else
         redirect_to(return_address)
       end
+    end
+
+    def fully_formed_url?(return_to)
+      uri = Addressable::URI.parse(return_to)
+      uri.present? && uri.scheme.present? && uri.host.present?
     end
 
     def decoded_host
